// src/app/enums/location-tags.ts
export enum LocationTags {
    wheelchairDifficult = "location.accessibility.DifficultForWheelchair", 
    disabledParking = "location.accessibility.DisabledParking",
    elevator = "location.accessibility.Elevator",
    inductiveHearingAid = "location.accessibility.InductiveHearingAid",
    infraredHearingAid = "location.accessibility.InfraredHearingAid",
    notForWheelchair = "location.accessibility.NotForWheelchair",
    radioHearingAid = "location.accessibility.RadioHearingAid",
    wcDin18024 = "location.accessibility.WC_DIN_18024",
    wheelchairAccessible = "location.accessibility.WheelchairAccessible"     
};